fix: server side using react props
