+++ src/app/tasks/[task_id]/video-details/VideoDetailsClient.tsx
@@ -5,33 +5,9 @@ import { useRouter } from 'next/navigation'
 import { ChevronLeft } from 'lucide-react'
 import Image from 'next/image'
 import type { TranscriptionResult } from '@/app/transcription/components/TranscriptionForm'
-import { gql } from '@apollo/client'
+import { gql, useMutation } from '@apollo/client'
 import client from '@/lib/apollo-client'
-import { AlertTriangle, ChevronUp, Mic } from 'lucide-react'
-
-const FIND_MULTIMEDIA_BY_TASK_ID = gql`
-  query FindMultimediaByTaskId($taskId: Int!) {
-    findMultimediaByTaskId(taskId: $taskId) {
-      id
-      taskId
-      photoUrl
-      videoUrl
-      audioTranscription
-    }
-  }
-`;
-
-function getMultimediaData(taskId: string) {
-  return client.query({
-    query: FIND_MULTIMEDIA_BY_TASK_ID,
-    variables: { taskId: Number(taskId) },
-  })
-    .then(({ data }) => data.findMultimediaByTaskId)
-    .catch(error => {
-      console.error("Error fetching multimedia data:", error);
-      return [];
-    });
-}
+import { AlertTriangle, ChevronUp, Mic, Trash2 } from 'lucide-react'
 
 interface MultimediaData {
   id: number;
@@ -41,25 +17,58 @@ interface MultimediaData {
   audioTranscription: string | null;
 }
 
-export default function VideoDetailsClient({ taskId }: { taskId: string }) {
+const DELETE_MULTIMEDIA = gql`
+  mutation DeleteMultimedia($deleteMultimediaId: Int!) {
+    deleteMultimedia(id: $deleteMultimediaId)
+  }
+`;
+
+export default function VideoDetailsClient({ taskId, initialVideoData }: { taskId: string, initialVideoData?: MultimediaData }) {
   const router = useRouter();
-  const [videoData, setVideoData] = useState<MultimediaData | undefined>(undefined);
-  const [isLoading, setIsLoading] = useState(true);
+  const [videoData] = useState<MultimediaData | undefined>(initialVideoData);
   const [transcriptionResult, setTranscriptionResult] = useState<TranscriptionResult | null>(null);
   const [loading, setLoading] = useState(true);
+  const [isDeleting, setIsDeleting] = useState(false);
+  const [showDeleteConfirm, setShowDeleteConfirm] = useState(false);
   
-  useEffect(() => {
-    getMultimediaData(taskId)
-      .then(multimediaData => {
-        const video = multimediaData.find((item: MultimediaData) => item.videoUrl);
-        setVideoData(video);
-        setIsLoading(false);
-      })
-      .catch(() => {
-        setIsLoading(false);
-      });
-  }, [taskId]);
+  // Initialize the delete mutation
+  const [deleteMultimedia] = useMutation(DELETE_MULTIMEDIA, {
+    client,
+    onCompleted: () => {
+      try {
+        localStorage.removeItem(`transcription-${taskId}`);
+        console.log(`Removed transcription-${taskId} from localStorage`);
+      } catch (error) {
+        console.error('Error removing item from localStorage:', error);
+      }
+      
+      router.push(`/tasks/${taskId}/risk_analysis`);
+    },
+    onError: (error) => {
+      console.error('Error deleting video:', error);
+      setIsDeleting(false);
+      setShowDeleteConfirm(false);
+    }
+  });
 
+  const handleDeleteVideo = async () => {
+    if (!videoData?.id) return;
+    
+    setIsDeleting(true);
+    try {
+      await deleteMultimedia({
+        variables: {
+          deleteMultimediaId: videoData.id
+        }
+      });
+      // The onCompleted callback will handle the redirect
+    } catch (error) {
+      console.error('Error in delete operation:', error);
+      setIsDeleting(false);
+      setShowDeleteConfirm(false);
+    }
+  };
+  
   useEffect(() => {
     if (videoData) {
       setTranscriptionResult({
@@ -90,7 +99,7 @@ export default function VideoDetailsClient({ taskId }: { taskId: string }) {
     }
   }, [taskId, videoData]);
 
-  if (isLoading || loading) {
+  if (loading) {
     return (
       <div className="min-h-screen bg-gray-100 flex items-center justify-center">
         <div className="animate-spin rounded-full h-12 w-12 border-t-2 border-b-2 border-teal-700"></div>
@@ -102,7 +111,7 @@ export default function VideoDetailsClient({ taskId }: { taskId: string }) {
     <div className="min-h-screen bg-gray-100 pb-6">
       <div className="bg-white p-4 shadow-sm">
         <button
-          onClick={() => router.back()}
+          onClick={() => router.push(`/tasks/${taskId}/risk_analysis`)}
           className="text-base text-red-500 mb-2 flex items-center font-medium"
         >
           <ChevronLeft className="h-5 w-5 mr-1" /> Volver
@@ -118,12 +127,24 @@ export default function VideoDetailsClient({ taskId }: { taskId: string }) {
           {/* Video preview */}
           <div className="relative aspect-video bg-gray-800 mb-6 rounded-md overflow-hidden">
             {videoData?.videoUrl ? (
-              <video 
-                src={videoData.videoUrl} 
-                controls 
-                className="w-full h-full object-cover"
-                poster="/video-thumbnail.jpg"
-              />
+              <>
+                <video 
+                  src={videoData.videoUrl} 
+                  controls 
+                  className="w-full h-full object-cover"
+                  poster="/video-thumbnail.jpg"
+                />
+                {/* Delete video button */}
+                {!isDeleting && (
+                  <button 
+                    onClick={() => setShowDeleteConfirm(true)}
+                    className="absolute top-2 right-2 bg-red-500 hover:bg-red-600 text-white p-2 rounded-full shadow-md transition-colors"
+                    title="Eliminar video"
+                  >
+                    <Trash2 className="h-5 w-5" />
+                  </button>
+                )}
+              </>
             ) : (
               <>
                 <div className="absolute inset-0 flex items-center justify-center">
@@ -145,6 +166,40 @@ export default function VideoDetailsClient({ taskId }: { taskId: string }) {
               </>
             )}
           </div>
+          
+          {/* Delete confirmation modal */}
+          {showDeleteConfirm && (
+            <div className="fixed inset-0 bg-black bg-opacity-50 z-50 flex items-center justify-center p-4">
+              <div className="bg-white rounded-lg p-6 max-w-sm w-full shadow-xl">
+                <h3 className="text-lg font-medium text-gray-900 mb-4">¿Estás seguro que deseas eliminar este video?</h3>
+                <p className="text-sm text-gray-500 mb-6">Esta acción no se puede deshacer. El video y su transcripción serán eliminados permanentemente.</p>
+                
+                <div className="flex justify-end space-x-3">
+                  <button
+                    onClick={() => setShowDeleteConfirm(false)}
+                    className="px-4 py-2 border border-gray-300 rounded-md text-gray-700 hover:bg-gray-50 font-medium"
+                    disabled={isDeleting}
+                  >
+                    Cancelar
+                  </button>
+                  <button
+                    onClick={handleDeleteVideo}
+                    className="px-4 py-2 bg-red-600 text-white rounded-md hover:bg-red-700 font-medium flex items-center"
+                    disabled={isDeleting}
+                  >
+                    {isDeleting ? (
+                      <>
+                        <div className="animate-spin rounded-full h-4 w-4 border-2 border-white border-t-transparent mr-2"></div>
+                        Eliminando...
+                      </>
+                    ) : (
+                      <>Eliminar video</>
+                    )}
+                  </button>
+                </div>
+              </div>
+            </div>
+          )}
           
           {/* Elements detected section */}
           <h2 className="text-lg font-semibold text-teal-700 mb-4">Elementos detectados</h2>
